+++ src/gui/status_bar.rs
@@ -1,7 +1,9 @@
-use strum::EnumMessage;
+use itertools::Itertools;
+use strum::{EnumMessage, IntoEnumIterator};
 
 use super::util::ResponseExt;
 use crate::app::App;
+use crate::puzzle::TwistMetric;
 
 pub fn build(ui: &mut egui::Ui, app: &mut App) {
     let mut changed = false;
@@ -22,15 +24,22 @@ pub fn build(ui: &mut egui::Ui, app: &mut App) {
         // Twist count
         let metric = &mut app.prefs.info.metric;
         let twist_count = app.puzzle.twist_count(*metric);
-        let r = ui
-            .add(
-                egui::Label::new(format!("{}: {}", metric, twist_count))
-                    .sense(egui::Sense::click()),
-            )
-            .on_hover_explanation(
+        let mut r = ui.add(
+            egui::Label::new(format!("{}: {}", metric, twist_count)).sense(egui::Sense::click()),
+        );
+        if ui.input().modifiers.shift {
+            r = r.on_hover_explanation(
+                "Twist count",
+                &TwistMetric::iter()
+                    .map(|m| format!("{m}: {}", app.puzzle.twist_count(m)))
+                    .join("\n"),
+            );
+        } else {
+            r = r.on_hover_explanation(
                 metric.get_message().unwrap_or(""),
                 metric.get_detailed_message().unwrap_or(""),
             );
+        }
         {
             let mut data = ui.data();
             let last_frame_metric = data.get_temp_mut_or_default(unique_id!());
@@ -42,8 +51,12 @@ pub fn build(ui: &mut egui::Ui, app: &mut App) {
                 ui.ctx().request_repaint();
             }
         }
-        if r.clicked() {
-            *metric = metric.next();
+        if r.clicked_by(egui::PointerButton::Primary) {
+            *metric = enum_iterator::next_cycle(metric).unwrap();
+            changed = true;
+        }
+        if r.clicked_by(egui::PointerButton::Secondary) {
+            *metric = enum_iterator::previous_cycle(metric).unwrap();
             changed = true;
         }
         ui.separator();